refactor: 코드리뷰 반영 useCallback을 통한 성능 최적화 및 getPageSize 함수 커스텀훅 내부로 이동
