+++ src/core/dropdown/Dropdown.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useCallback, useState } from "react";
 import useResize from "../../lib/hooks/useResize";
 
 import ArrowDown from "../assets/icons/dropdown/arrow_down.svg";
@@ -13,13 +13,7 @@ interface DropdownProps {
 }
 
 const Dropdown = ({ isLoading, order, handleListClick }: DropdownProps) => {
-  const getPageSize = (width: number) => {
-    if (width < 768) {
-      return 1;
-    }
-    return 0;
-  };
-  const { size } = useResize({ getPageSize });
+  const { size } = useResize({ pcSize: 0, tabletSize: 0, mobileSize: 1 });
   const [isDrop, setIsDrop] = useState<boolean>(false);
   const handleDropdown = () => {
     setIsDrop(!isDrop);
